+++ component/file_cache/file_cache.go
@@ -191,7 +191,7 @@ func (fc *FileCache) GenConfig() string {
 	log.Info("FileCache::Configure : config generation started")
 
 	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("\n%s:", fc.Name()))
+	fmt.Fprintf(&sb, "\n%s:", fc.Name())
 
 	tmpPath := ""
 	_ = config.UnmarshalKey("tmp-path", &tmpPath)
@@ -204,8 +204,8 @@ func (fc *FileCache) GenConfig() string {
 		timeout = 0
 	}
 
-	sb.WriteString(fmt.Sprintf("\n  path: %v", common.ExpandPath(tmpPath)))
-	sb.WriteString(fmt.Sprintf("\n  timeout-sec: %v", timeout))
+	fmt.Fprintf(&sb, "\n  path: %v", common.ExpandPath(tmpPath))
+	fmt.Fprintf(&sb, "\n  timeout-sec: %v", timeout)
 
 	return sb.String()
 }
@@ -993,9 +993,9 @@ func (fc *FileCache) OpenFile(options internal.OpenFileOptions) (*handlemap.Hand
 				// File was created locally and now download has failed so we need to delete it back from local cache
 				log.Err("FileCache::OpenFile : error downloading file from storage %s [%s]", options.Name, err.Error())
 				_ = f.Close()
-				err = os.Remove(localPath)
-				if err != nil {
-					log.Err("FileCache::OpenFile : Failed to remove file %s [%s]", localPath, err.Error())
+				err2 := os.Remove(localPath)
+				if err2 != nil {
+					log.Err("FileCache::OpenFile : Failed to remove file %s [%s]", localPath, err2.Error())
 				}
 				return nil, err
 			}