+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 2.5.3 (Unreleased)
 **Features**
+- Add rate limit functionality for ingress bandwidth (bytes downloaded per second) and operations per second ([PR #2093](https://github.com/Azure/azure-storage-fuse/pull/2093))
 
 **Bug Fixes**
+- Fix panic while reading an archived blob using file-cache ([PR #2127](https://github.com/Azure/azure-storage-fuse/pull/2127)]
+- Ensure logger is properly deinitialized on mount failures ([PR #2125](https://github.com/Azure/azure-storage-fuse/pull/2125)]
 
 ## 2.5.2 (2026-01-19)
 **Features**